split utxo from transaction
